Rename tictactoe.rb into game_run.rb
