+++ lib/helpers/game_run.rb
