+++ lib/main.rb
@@ -1,4 +1,4 @@
-require_relative "./helpers/tic_tac_toe.rb"
+require_relative "./helpers/game_run.rb"
 require_relative "./helpers/assets.rb"
 
 show_title
